refactor: like product function
